+++ spec/integrations/instrumentation/vendors/kubernetes/polling_liveness_timeout_flow_spec.rb
@@ -32,6 +32,10 @@ Thread.new do
     uri = URI.parse('http://127.0.0.1:9005/')
     response = Net::HTTP.get_response(uri)
     DT[:probing] << response.code
+
+    next if response.code == '200'
+
+    DT[:bodies] << response.body
   end
 end
 
@@ -43,5 +47,15 @@ start_karafka_and_wait_until do
   DT.key?(0)
 end
 
-assert DT[:probing].include?('204')
+assert DT[:probing].include?('200')
 assert DT[:probing].include?('500')
+
+last = JSON.parse(DT[:bodies].last)
+
+assert_equal 'unhealthy', last['status']
+assert last.key?('timestamp')
+assert_equal 9005, last['port']
+assert_equal Process.pid, last['process_id']
+assert_equal true, last['errors']['polling_ttl_exceeded']
+assert_equal false, last['errors']['consumption_ttl_exceeded']
+assert_equal false, last['errors']['unrecoverable']